fix: add new hide_presentation_on_join flag
As of BBB 2.6 the userdata flag for hiding the presentation has changed. This commit adds the new `bbb_hide_presentation_on_join` flag when a clean layout is selected.

In order to keep backwards compatability for BBB 2.5 and lower we can
send both flags and remove the deprecated one later on.